task 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,7 +10,7 @@
 
 int main(void)
 {
-	unsigned long int a, b, sum;
+	unsigned long long int a, b, sum;
 	int count;
 
 	a = 0;
@@ -21,12 +21,12 @@ int main(void)
 	while (count < 98)
 	{
 		sum = a + b;
-		printf("%lu, ", sum);
+		printf("%d, ", sum);
 		a = b;
 		b = sum;
 		count++;
 	}
 
-	printf("%lu\n", sum);
+	printf("%d\n", sum);
 	return (0);
 }
